requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
bcrypt==5.0.0
pytest==9.0.3
pytest-django==4.12.0